Adds: backward compatibility of root file to RDF for reprocessing

+++ DreamAN/Cuts/EventCut.h
@@ -23,8 +23,8 @@ class EventCut {
   static EventCut* PhotonCuts();
   // Filter function for RDataFrame
   bool operator()(const std::vector<int>& pid, const std::vector<float>& px, const std::vector<float>& py, const std::vector<float>& pz, const std::vector<float>& vx,
-                  const std::vector<float>& vy, const std::vector<float>& vz, const std::vector<float>& vt, const std::vector<int>& charge, const std::vector<float>& beta,
-                  const std::vector<float>& chi2pid, const std::vector<int>& status, const std::vector<int>& REC_Track_pass_fid) const;
+                  const std::vector<float>& vy, const std::vector<float>& vz, const std::vector<float>& vt, const std::vector<short>& charge, const std::vector<float>& beta,
+                  const std::vector<float>& chi2pid, const std::vector<short>& status, const std::vector<int>& REC_Track_pass_fid) const;
 
  private:
   // Filtering condition ranges
